requirements.txt
langchain
streamlit
python-dotenv
PyPDF2
pandas
langchain-google-genai
google-genai
langchain-community
python-dotenv

-e .